backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
pdfplumber==0.7.5
requests==2.28.2
gunicorn==20.1.0
python-dotenv==0.21.0
